# Conflicts:
#	db.sqlite3

+++ mainsite/models.py
@@ -6,7 +6,7 @@ from django.db import models
 class History(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
     wechat_id = models.CharField(max_length=100, name='微信号')
-    number = models.IntegerField(name="数字")
+    number = models.FloatField(name="数字")
     grade = models.PositiveSmallIntegerField(name='年级')
     class_number = models.PositiveSmallIntegerField(name='班级')
 
@@ -14,14 +14,14 @@ class History(models.Model):
 class Winners(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
     wechat_id = models.CharField(max_length=100, name='微信号')
-    number = models.IntegerField(name="数字")
+    number = models.FloatField(name="数字")
     grade = models.PositiveSmallIntegerField(name='年级')
     class_number = models.PositiveSmallIntegerField(name='班级')
     delta = models.PositiveSmallIntegerField()
 
 
 class Number(models.Model):
-    number = models.PositiveSmallIntegerField()
+    number = models.FloatField()
     in_progress = models.BooleanField()
 
     def __str__(self):
